shell: merge config change handlers
This combines the config change handlers into a single function.
